requirements.txt
streamlit==1.35.0
mysql-connector-python==8.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scikit-fuzzy==0.4.2
matplotlib==3.9.0
plotly==5.22.0
bcrypt==4.1.3
Pillow==10.3.0
openpyxl==3.1.4
xlrd==2.0.1
nltk==3.8.1
PySastrawi==1.2.0
python-dotenv==1.0.1
streamlit-option-menu==0.3.12
streamlit-aggrid==0.3.4.post3
extra-streamlit-components==0.1.71
